pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-estado-peru"
version = "0.1.0"
description = "Sistema RAG para consultas sobre normativa pública peruana"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ruben"}
]
keywords = ["rag", "llm", "gemini", "peru", "legal"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    # LLM Providers
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",  # Para Groq (usa API compatible con OpenAI)
    # Vector Store & Embeddings
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    # Document Processing
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.3",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
]
eval = [
    "ragas>=0.1.0",
    "datasets>=2.16.0",
]

[project.scripts]
rag-ingest = "scripts.ingest:main"
rag-query = "scripts.query:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["packages*", "services*", "scripts*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
